--- src/pages/admin/Collection/Page.tsx
@@ -1,7 +1,0 @@
-export default function Page() {
-  return (
-    <div>
-      <h1>Hello It's Collection</h1>
-    </div>
-  );
-}
