+++ src/pages/members/MembershipStore/MembershipStore.tsx
@@ -1,4 +1,4 @@
-export default function Page() {
+export default function MembershipStore() {
   return (
     <div>
       <h1>Hello It's MembershipStore</h1>
