+++ src/pages/members/MyMembership/MyMembership.tsx
@@ -1,4 +1,4 @@
-export default function Page() {
+export default function MyMembership() {
   return (
     <div>
       <h1>Hello It's MyMembership</h1>
